More work on MockHttp stuff
